backend/pyproject.toml
[project]
name = "Vectora_Project_Backend"
version = "0.1.0"
description = "An AI-powered personal content vault that lets you save images, documents, and notes, search them instantly with semantic intelligence, and chat with your content for quick insights — all with privacy at the core."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-sam-cli>=1.148.0",
    "boto3>=1.42.35",
    "cohere>=5.20.1",
    "easyocr>=1.7.2",
    "fastapi[standard]>=0.124.0",
    "firebase-admin>=7.1.0",
    "google-genai>=1.56.0",
    "langchain>=1.2.3",
    "langchain-google-genai>=4.1.3",
    "langchain-openai>=1.1.7",
    "langchain-pinecone>=0.2.13",
    "mangum>=0.20.0",
    "pillow>=12.0.0",
    "pyjwt>=2.10.1",
    "pymupdf>=1.26.7",
    "python-multipart>=0.0.20",
    "razorpay>=2.0.0",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "voyageai>=0.3.7",
]
